requirements.txt
google_api_python_client==2.81.0
google_auth_oauthlib==1.0.0
protobuf==4.22.1
pyTelegramBotAPI==4.10.0
pytz==2022.7.1
